requirements.txt
django
djangorestframework
mysqlclient
requests 